Update testGame.cpp
pieces now look to see if they will put the king in check when the piece goes to move. This means that you must have a king of that color when ever you call a move method! otherwise you get a null pointer error because we are asking for a king that doesn't exist